Initial import of Unique
Unique is a library for writing single instance applications. An application
using Unique will block any othe instance of the same application launched
after the first one; any other instance can send a message to the running
application, for instance to open a list of URIs.

Unique can be compiled against different backends, implementing the IPC
mechanisms used by different instances to communicate.

This is the initial import of the Unique library: at the moment, only the
(default) D-Bus backend is working.

The public API is not yet frozen.