Remove screen argument from the message data API
The GdkScreen is automatically set up by Unique when sending a message
to the running instance, as well as the startup notification id.

Allow passing a NULL UniqueMessageData to unique_app_send_message(): the
message will be created instead of copied.

Fix a couple of leaks while we are at it.