Add documentation and fallback to unique_app_send_message()
Now unique_app_send_message() checks whether is-running is FALSE and
returns UNIQUE_RESPONSE_INVALID if it's the case. I do not have strong
opinions on this one, though; it might make sense to actually emit the
message-received signal on the same instance, but it would be a programming
error to call unique_app_new() and unique_app_send_message() on the running
instance without actually checking with unique_app_is_running() beforehand.