Provide a default handler for ::message-received
Without a default handler, the message accumulation would fall
through and the response would always be INVALID.

The intended behaviour is that the ::message-received signal
emission chain should continue until a handler returns a valid
response code meaning: "I handled the signal". What is missing
is a way to tell the signal machinery that the handler did not
handle the signal. Hence, the UNIQUE_RESPONSE_PASSTHROUGH
response code has been added.

The signal emission chain also needs a final handler, some place
that says that the bucket stops there; for this, we need a
class handler for the ::message-received signal. The default
behaviour, if no handler returned a valid code or if no handler
has been set, is to just call gtk_window_present() on the first
window inside the watch list. Obviously, if this is not the
wanted behaviour, you can either connect a handler to the signal
and stop the emission chain or you can subclass the UniqueApp
class and override the class handler yourself.

Thanks to John Wendell for pointing out this gaping bug and for
the investigation.