[message] Add raw data getter
UniqueMessageData has a setter function for setting raw data inside
the structure, but it is missing a raw data getter.

This commit adds unique_message_data_get() which returns a pointer to
the internal raw data and the length of the data.

Using this new getter it is possible to implement other types of accessor
functions than the text/URIs LibUnique already ships.