[message] Add filename accessors
Allow setting and getting a filename inside a UniqueMessageData.

Using set_text()/get_text() to set filenames might break on non-UTF-8
locales because of the implicit text conversion; for this we need an
accessor pair that sets a raw C string. Since we added the raw getter
in a previous commit, it's safe to provide a simple wrapper around
unique_message_data_set()/unique_message_data_get() for setting plain
C strings such as filenames.

Kudos to Alex Larsson for the idea.