Set the enumeration values as static
Apparently, when registering an enumeration GType, you have to
declar the GEnumValue array holding value, nick and name as a
"static const" instead of just "static". This is quite counter-
intuitive, since we already have a static variable posing as a
guard in place inside the _get_type() function. The GEnum
documentation is also particularly scarce in this regard.

The fix is trivial and does not have side effects, and as it makes
libunique behave correctly on OpenBSD so it's a good thing to have
in the main tree.

Patch by: Landry Breuil <gaston@gcu.info>

Signed-off by: Emmanuele Bassi <ebassi@linux.intel.com>