Do not attempt to convert UTF-8 data
Bug 575826 – libunique unique_message_data_[get|set]_text don't work
             on none UTF-8

When unique_message_data_set_text() is called on non UTF-8 locales,
the argument message_data->data is always set to UTF-8.

And when unique_message_data_get_text() is called on non UTF-8 locales,
it tries to convert the current encoding to UTF-8 even if the string is
already UTF-8, so the conversion fails

The data should be validated and converted only if the UTF-8 validation
fails.

Signed-off-by: Emmanuele Bassi <ebassi@gnome.org>