dbus: Set default timeout to 3 seconds
1. Start a unique app (say, empathy)
2. Make it unresponsive (SIGSTOP is the easy way)
3. Start it again
4. ... 30 seconds passes ...

(empathy:9853): Unique-DBus-WARNING **: Error while sending message: Did
not receive a reply. Possible causes include: the remote application did
not send a reply, the message bus security policy blocked the reply, the
reply timeout expired, or the network connection was broken.

The timeout really should be tamped down here, perhaps to a couple of
seconds... realistically, if the app isn't responding at that point,
it's not going to.

https://bugzilla.gnome.org/show_bug.cgi?id=629781

Signed-off-by: Emmanuele Bassi <ebassi@linux.intel.com>